+++ changelog.md
@@ -7,6 +7,7 @@
 - /record command to generate link for recording game
 - recording from an lfg thread helps pre-build the game form
 - series can now reorganize tabs and any tab can be hidden
+- added elo ranking as an optional tab to series
 
 ### Improvements
 - /random commands now have option to include fan content